Remove optimized js dir from .gitignore
The `make build` step from the last commit requires node, so rather than
require that on the deployment image, I'm choosing to build and version
control the optimized js dir. So the `make build` step is removed from
the init script and the optimized js dir is added to the repo.

Signed-off-by: Aaron Browne <aaron0browne@gmail.com>